@@ -12,7 +12,7 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">7.1.3",
+        "php": ">=7.4",
         "ext-gmp": "*",
         "ext-json": "*",
         "ext-mbstring": "*",
