@@ -17,12 +17,12 @@ class UserAuthenticationService
     /**
      * @var \Aws\CognitoIdentityProvider\CognitoIdentityProviderClient
      */
-    private $cognitoClient;
+    private CognitoClient $cognitoClient;
 
     /**
      * @var \Incognito\CognitoClient\CognitoCredentials
      */
-    private $cognitoCredentials;
+    private CognitoCredentials $cognitoCredentials;
 
     /**
      * Constructor.
